rearrange the structure of the classes. Configuration reading and scanning classes are divided into two.  API builder to scan classes and build API definition.
