+++ modules/api-everywhere-as/src/main/java/org/wso2/appserver/apieverywhere/APIEverywhereInitiator.java
@@ -4,9 +4,11 @@ package org.wso2.appserver.apieverywhere;
 import org.apache.juli.logging.Log;
 import org.apache.juli.logging.LogFactory;
 import org.wso2.appserver.apieverywhere.exceptions.APIEverywhereException;
+import org.wso2.appserver.apieverywhere.utils.APICreateRequest;
 import org.wso2.appserver.configuration.context.WebAppApiEverywhere;
 import org.wso2.appserver.configuration.listeners.ContextConfigurationLoader;
 
+import java.util.HashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import javax.servlet.ServletContext;
@@ -24,8 +26,10 @@ import javax.servlet.annotation.WebListener;
 public class APIEverywhereInitiator implements ServletContextListener {
 
     private static final Log log = LogFactory.getLog(APIEverywhereInitiator.class);
-    private APIScanner apiScanner = new APIScanner();
+    private ConfigScanner configScanner = new ConfigScanner();
+    private APIBuilder apiBuilder = new APIBuilder();
     private APICreator apiCreator = new APICreator();
+    private APICreateRequest apiCreateRequest = new APICreateRequest();
     private ExecutorService executor = Executors.newSingleThreadExecutor();
 
     //catch a web app deployment event
@@ -38,10 +42,15 @@ public class APIEverywhereInitiator implements ServletContextListener {
 
                     if (apiEverywhereConfiguration != null && apiEverywhereConfiguration.getCreateApi() != null) {
                         if (apiEverywhereConfiguration.getCreateApi()) {
-                            apiScanner.scan(servletContext).ifPresent(apiRequest -> {
+                            HashMap<String, StringBuilder> beanParams = configScanner.scanConfigs(servletContext);
+                            apiCreateRequest.setContext(servletContext.getContextPath());
+                            apiCreateRequest.setName(servletContext.getContextPath().substring(1));
+                            if (!beanParams.isEmpty()) {
+                                String apiRequest = apiBuilder.build(beanParams, apiCreateRequest);
                                 apiCreator.addAPIRequest(apiRequest);
                                 executor.execute(apiCreator);
-                            });
+                                executor.shutdown();
+                            }
                         } else {
                             if (log.isDebugEnabled()) {
                                 log.debug("Creation of API is blocked by the user.");
